Replace Navy file with local Skyfield usage
The Navy website has been down for a while now, and while it was nice to have
results computed in year chunks, I still want it to function, so now run the
computations locally with this astronomy library. It will automatically
download data files it needs, but those seem more reliable than a Navy-hosted
Perl script.